Fix memory and type overflow bugs in UDP traffic generator
- Fix type mismatch: Change tokenBucket from unsigned long to long long
- Add input validation: Maximum CBR limit of 500 Gbps (500,000 Mbps)
- Add overflow protection: Check before multiplication in token bucket
- Fix Makefile: Use gcc for C files, g++ for C++ files to resolve linking

These changes prevent undefined behavior, integer overflow, and improve
input validation for network testing at high bit rates.

+++ udpCbrDriver.h
@@ -1,1 +1,1 @@
-extern unsigned long tokenBucket;
+extern long long tokenBucket;
